+++ tests/integration/dlc/dlc.test.ts
@@ -48,8 +48,6 @@ import dlcTxsJSON from '../fixtures/dlctxs.json';
 import Oracle from '../models/Oracle';
 import {
   generateContractInfo,
-  generateContractInfoCustomStrategyOracle18,
-  generateContractInfoCustomStrategyOracle27,
   generateOracleAttestation,
 } from '../utils/contract';
 
@@ -102,290 +100,6 @@ describe('inputToFundingInput', () => {
     expect(
       alice.dlc.inputToFundingInput(invalidInput),
     ).to.be.eventually.rejectedWith(Error);
-  });
-});
-
-describe.only('Custom Strategy Oracle POC numdigits=27', () => {
-  const numDigits = 27;
-  const oracleBase = 2;
-  const eventId = 'strategyOutcome';
-  let dlcOffer: DlcOffer;
-  let dlcAccept: DlcAccept;
-  let dlcSign: DlcSign;
-  let dlcTransactions: DlcTransactions;
-  let oracleAttestation: OracleAttestationV0;
-  let aliceAddresses: string[] = [];
-  let oracle: Oracle;
-
-  before(async () => {
-    const aliceNonChangeAddresses: string[] = (
-      await alice.wallet.getAddresses(0, 15, false)
-    ).map((address) => address.address);
-    const aliceChangeAddresses: string[] = (
-      await alice.wallet.getAddresses(0, 15, true)
-    ).map((address) => address.address);
-
-    aliceAddresses = [...aliceNonChangeAddresses, ...aliceChangeAddresses];
-
-    for (let i = 0; i < aliceAddresses.length; i++) {
-      await alice.getMethod('jsonrpc')(
-        'importaddress',
-        aliceAddresses[i],
-        '',
-        false,
-      );
-    }
-  });
-
-  it('should complete entire flow', async () => {
-    const aliceInput = await getInput(alice);
-    const bobInput = await getInput(bob);
-
-    oracle = new Oracle('olivia', numDigits);
-
-    const {
-      contractInfo,
-      totalCollateral,
-    } = generateContractInfoCustomStrategyOracle27(
-      oracle,
-      numDigits,
-      oracleBase,
-    );
-
-    const feeRatePerVb = BigInt(10);
-    const cetLocktime = 1617170572;
-    const refundLocktime = 1617170573;
-    const premium = BigInt(50000);
-    const outcome = 105000000;
-
-    console.time('total');
-    console.time('offer');
-
-    dlcOffer = await alice.dlc.createDlcOffer(
-      contractInfo,
-      totalCollateral - premium,
-      feeRatePerVb,
-      cetLocktime,
-      refundLocktime,
-      [aliceInput],
-    );
-
-    console.timeEnd('offer');
-
-    console.time('accept');
-
-    const acceptDlcOfferResponse: AcceptDlcOfferResponse = await bob.dlc.acceptDlcOffer(
-      dlcOffer,
-      [bobInput],
-    );
-    dlcAccept = acceptDlcOfferResponse.dlcAccept;
-    dlcTransactions = acceptDlcOfferResponse.dlcTransactions;
-
-    console.timeEnd('accept');
-
-    const { dlcTransactions: dlcTxsFromMsgs } = await bob.dlc.createDlcTxs(
-      dlcOffer,
-      dlcAccept,
-    );
-
-    console.time('sign');
-
-    const signDlcAcceptResponse: SignDlcAcceptResponse = await alice.dlc.signDlcAccept(
-      dlcOffer,
-      dlcAccept,
-    );
-    console.timeEnd('sign');
-
-    dlcSign = signDlcAcceptResponse.dlcSign;
-
-    console.time('finalize');
-
-    const fundTx = await bob.dlc.finalizeDlcSign(
-      dlcOffer,
-      dlcAccept,
-      dlcSign,
-      dlcTransactions,
-    );
-
-    console.timeEnd('finalize');
-
-    const fundTxId = await bob.chain.sendRawTransaction(
-      fundTx.serialize().toString('hex'),
-    );
-
-    console.time('attestation');
-    oracleAttestation = generateOracleAttestation(
-      outcome,
-      oracle,
-      oracleBase,
-      numDigits,
-      eventId,
-    );
-    console.timeEnd('attestation');
-
-    console.time('execute');
-
-    const cet = await bob.dlc.execute(
-      dlcOffer,
-      dlcAccept,
-      dlcSign,
-      dlcTransactions,
-      oracleAttestation,
-      false,
-    );
-
-    console.timeEnd('execute');
-    console.timeEnd('total');
-
-    console.log('CET Outputs: ', cet.toJSON().outputs);
-
-    const cetTxId = await bob.chain.sendRawTransaction(
-      cet.serialize().toString('hex'),
-    );
-    const cetTx = await alice.getMethod('getTransactionByHash')(cetTxId);
-    expect(cetTx._raw.vin.length).to.equal(1);
-  });
-});
-
-describe.only('Custom Strategy Oracle POC numdigits=18', () => {
-  const numDigits = 18;
-  const oracleBase = 2;
-  const eventId = 'strategyOutcome';
-  let dlcOffer: DlcOffer;
-  let dlcAccept: DlcAccept;
-  let dlcSign: DlcSign;
-  let dlcTransactions: DlcTransactions;
-  let oracleAttestation: OracleAttestationV0;
-  let aliceAddresses: string[] = [];
-  let oracle: Oracle;
-
-  before(async () => {
-    const aliceNonChangeAddresses: string[] = (
-      await alice.wallet.getAddresses(0, 15, false)
-    ).map((address) => address.address);
-    const aliceChangeAddresses: string[] = (
-      await alice.wallet.getAddresses(0, 15, true)
-    ).map((address) => address.address);
-
-    aliceAddresses = [...aliceNonChangeAddresses, ...aliceChangeAddresses];
-
-    for (let i = 0; i < aliceAddresses.length; i++) {
-      await alice.getMethod('jsonrpc')(
-        'importaddress',
-        aliceAddresses[i],
-        '',
-        false,
-      );
-    }
-  });
-
-  it('should complete entire flow', async () => {
-    const aliceInput = await getInput(alice);
-    const bobInput = await getInput(bob);
-
-    oracle = new Oracle('olivia', numDigits);
-
-    const {
-      contractInfo,
-      totalCollateral,
-    } = generateContractInfoCustomStrategyOracle18(
-      oracle,
-      numDigits,
-      oracleBase,
-    );
-
-    const feeRatePerVb = BigInt(10);
-    const cetLocktime = 1617170572;
-    const refundLocktime = 1617170573;
-    const premium = BigInt(5);
-    const outcome = 10500;
-
-    console.time('total');
-    console.time('offer');
-
-    dlcOffer = await alice.dlc.createDlcOffer(
-      contractInfo,
-      totalCollateral - premium,
-      feeRatePerVb,
-      cetLocktime,
-      refundLocktime,
-      [aliceInput],
-    );
-
-    console.timeEnd('offer');
-
-    console.time('accept');
-
-    const acceptDlcOfferResponse: AcceptDlcOfferResponse = await bob.dlc.acceptDlcOffer(
-      dlcOffer,
-      [bobInput],
-    );
-    dlcAccept = acceptDlcOfferResponse.dlcAccept;
-    dlcTransactions = acceptDlcOfferResponse.dlcTransactions;
-
-    console.timeEnd('accept');
-
-    const { dlcTransactions: dlcTxsFromMsgs } = await bob.dlc.createDlcTxs(
-      dlcOffer,
-      dlcAccept,
-    );
-
-    console.time('sign');
-
-    const signDlcAcceptResponse: SignDlcAcceptResponse = await alice.dlc.signDlcAccept(
-      dlcOffer,
-      dlcAccept,
-    );
-    console.timeEnd('sign');
-
-    dlcSign = signDlcAcceptResponse.dlcSign;
-
-    console.time('finalize');
-
-    const fundTx = await bob.dlc.finalizeDlcSign(
-      dlcOffer,
-      dlcAccept,
-      dlcSign,
-      dlcTransactions,
-    );
-
-    console.timeEnd('finalize');
-
-    const fundTxId = await bob.chain.sendRawTransaction(
-      fundTx.serialize().toString('hex'),
-    );
-
-    console.time('attestation');
-    oracleAttestation = generateOracleAttestation(
-      outcome,
-      oracle,
-      oracleBase,
-      numDigits,
-      eventId,
-    );
-    console.timeEnd('attestation');
-
-    console.time('execute');
-
-    const cet = await bob.dlc.execute(
-      dlcOffer,
-      dlcAccept,
-      dlcSign,
-      dlcTransactions,
-      oracleAttestation,
-      false,
-    );
-
-    console.timeEnd('execute');
-    console.timeEnd('total');
-
-    console.log('CET Outputs: ', cet.toJSON().outputs);
-
-    const cetTxId = await bob.chain.sendRawTransaction(
-      cet.serialize().toString('hex'),
-    );
-    const cetTx = await alice.getMethod('getTransactionByHash')(cetTxId);
-    expect(cetTx._raw.vin.length).to.equal(1);
   });
 });
