+++ tests/integration/utils/contract.ts
@@ -1,4 +1,4 @@
-import { CoveredCall, LinearPayout } from '@node-dlc/core';
+import { CoveredCall } from '@node-dlc/core';
 import {
   ContractDescriptorV1,
   ContractInfoV0,
@@ -7,6 +7,7 @@ import {
   OracleAttestationV0,
   OracleEventV0,
   OracleInfoV0,
+  PayoutFunctionV0,
   RoundingIntervalsV0,
 } from '@node-dlc/messaging';
 import { math } from 'bip-schnorr';
@@ -76,10 +77,14 @@ export function generateContractInfo(
   return { contractInfo, totalCollateral };
 }
 
-export function generateContractInfoCustomStrategyOracle27(
+export function generateContractInfoCustomStrategyOracle(
   oracle: Oracle,
   numDigits = 18,
   oracleBase = 2,
+  payoutFunction: PayoutFunctionV0,
+  intervals: { beginInterval: bigint; roundingMod: bigint }[],
+  totalCollateral: bigint,
+  unit = 'BTC',
   eventId = 'strategyOutcome',
 ): { contractInfo: ContractInfoV0; totalCollateral: bigint } {
   const oliviaInfo = oracle.GetOracleInfo();
@@ -87,7 +92,7 @@ export function generateContractInfoCustomStrategyOracle27(
   const eventDescriptor = new DigitDecompositionEventDescriptorV0();
   eventDescriptor.base = oracleBase;
   eventDescriptor.isSigned = false;
-  eventDescriptor.unit = 'BTC';
+  eventDescriptor.unit = unit;
   eventDescriptor.precision = 0;
   eventDescriptor.nbDigits = numDigits;
 
@@ -115,89 +120,6 @@ export function generateContractInfoCustomStrategyOracle27(
   const oracleInfo = new OracleInfoV0();
   oracleInfo.announcement = announcement;
 
-  // Assumption that min and max payout are 0.8 BTC and 1.2 BTC respectively
-  const { payoutFunction } = LinearPayout.buildPayoutFunction(
-    80000000n,
-    120000000n,
-    80000000n,
-    120000000n,
-    oracleBase,
-    numDigits,
-  );
-
-  // Assumption that total collateral is always 1 BTC
-  const totalCollateral = 120000000n;
-
-  const intervals = [{ beginInterval: 0n, roundingMod: 50000n }];
-  const roundingIntervals = new RoundingIntervalsV0();
-  roundingIntervals.intervals = intervals;
-
-  const contractDescriptor = new ContractDescriptorV1();
-  contractDescriptor.numDigits = numDigits;
-  contractDescriptor.payoutFunction = payoutFunction;
-  contractDescriptor.roundingIntervals = roundingIntervals;
-
-  const contractInfo = new ContractInfoV0();
-  contractInfo.totalCollateral = totalCollateral;
-  contractInfo.contractDescriptor = contractDescriptor;
-  contractInfo.oracleInfo = oracleInfo;
-
-  return { contractInfo, totalCollateral };
-}
-
-export function generateContractInfoCustomStrategyOracle18(
-  oracle: Oracle,
-  numDigits = 18,
-  oracleBase = 2,
-  eventId = 'strategyOutcome',
-): { contractInfo: ContractInfoV0; totalCollateral: bigint } {
-  const oliviaInfo = oracle.GetOracleInfo();
-
-  const eventDescriptor = new DigitDecompositionEventDescriptorV0();
-  eventDescriptor.base = oracleBase;
-  eventDescriptor.isSigned = false;
-  eventDescriptor.unit = 'BTC';
-  eventDescriptor.precision = 0;
-  eventDescriptor.nbDigits = numDigits;
-
-  const event = new OracleEventV0();
-  event.oracleNonces = oliviaInfo.rValues.map((rValue) =>
-    Buffer.from(rValue, 'hex'),
-  );
-  event.eventMaturityEpoch = 1617170572;
-  event.eventDescriptor = eventDescriptor;
-  event.eventId = eventId;
-
-  const announcement = new OracleAnnouncementV0();
-  announcement.announcementSig = Buffer.from(
-    oracle.GetSignature(
-      math
-        .taggedHash('DLC/oracle/announcement/v0', event.serialize())
-        .toString('hex'),
-    ),
-    'hex',
-  );
-
-  announcement.oraclePubkey = Buffer.from(oliviaInfo.publicKey, 'hex');
-  announcement.oracleEvent = event;
-
-  const oracleInfo = new OracleInfoV0();
-  oracleInfo.announcement = announcement;
-
-  // Assumption that min and max payout are 0.8 BTC and 1.2 BTC respectively
-  const { payoutFunction } = LinearPayout.buildPayoutFunction(
-    8000n,
-    12000n,
-    8000n,
-    12000n,
-    oracleBase,
-    numDigits,
-  );
-
-  // Assumption that total collateral is always 1 BTC
-  const totalCollateral = 12000n;
-
-  const intervals = [{ beginInterval: 0n, roundingMod: 5n }];
   const roundingIntervals = new RoundingIntervalsV0();
   roundingIntervals.intervals = intervals;
